% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_req.R
\name{get_rand_quote}
\alias{get_rand_quote}
\title{get_rand_quote}
\usage{
get_rand_quote(n = 1, perso = NULL)
}
\arguments{
\item{n}{nombre de citations}

\item{perso}{OPTIONNEL, spécifier un nom de pesonnage, ou plusieurs avec un vecteur}
}
\description{
Génère n citations Kaamelott formattées pour terminal
}
